requirements.txt
# Core
fastapi==0.110.0
uvicorn[standard]==0.29.0
streamlit==1.32.0

# Data
pandas==2.2.1
numpy==1.26.4
duckdb==0.10.1
pyarrow==15.0.0

# ML
prophet==1.1.5
scikit-learn==1.4.1

# Scheduling
APScheduler==3.10.4

# Validation
pandera==0.18.3

# Visualization
plotly==5.20.0

# Alerting
httpx==0.27.0

# Testing
pytest==8.1.1
pytest-cov==5.0.0
httpx==0.27.0

# Linting
ruff==0.3.4

# Utils
python-dotenv==1.0.1
pydantic==2.6.4
pydantic-settings==2.2.1
